+++ libraries/AtomicWeight/elements_float.h
@@ -140,4 +140,5 @@ elements[119] =
 };
 
 
-// -- END OF FILE --
+//  -- END OF FILE --
+